requirements.txt
Django==5.2.7
gunicorn==23.0.0
uvicorn==0.32.1
psycopg2-binary==2.9.9
Pillow==11.1.0
django-taggit==6.1.0
django-ckeditor-5==0.2.16
django-redis==5.4.0
whitenoise==6.8.2

# AI & Vector Support
ollama==0.4.7
django-environ==0.11.2
numpy==2.2.1
redis>=5.0.0
